fix(manage-components): Add confirmation dialog for PSU deletion action

+++ admin/manage_components.php
@@ -271,7 +271,7 @@ if (hesk_SESSION('iserror')) {
                                     <div class="actions">
                                         <a class="action-btn edit" href="manage_component.php?do=edit&type=psu&id=<?php echo $psu['id'] ?>"><svg class="icon icon-edit-ticket"><use xlink:href="../img/sprite.svg#icon-edit-ticket"></use></svg></a>
                                         <a class="action-btn view" href="manage_component.php?do=view&type=psu&id=<?php echo $psu['id'] ?>"><svg class="icon icon-view"><use xlink:href="../img/sprite.svg#icon-eye-open"></use></svg></a>
-                                        <a class="action-btn delete" href="manage_component.php?do=delete&type=psu&id=<?php echo $psu['id'] ?>"><svg class="icon icon-delete"><use xlink:href="../img/sprite.svg#icon-delete"></use></svg></a>                                
+                                        <a class="action-btn delete" href="manage_component.php?do=delete&type=psu&id=<?php echo $psu['id'] ?>" onclick="return confirm('<?php echo $hesklang['delete_confirm'].' '.$psu['model'].'?' ?>')"><svg class="icon icon-delete"><use xlink:href="../img/sprite.svg#icon-delete"></use></svg></a>                                
                                     </div>
                                 </td>
                             </tr>
